Move BRIDGE variable definition to standalone network functions
The standalone networking functions are the only places that use the
BRIDGE environment variable. Right now, it's defined in
deploy-standalone, which needs to be run after one of the two networking
functions are called. Since the networking functions lay down the
standalone_parameters, the parameters are incomplete since the BRIDGE
variable is empty, cause the deploy-standalone function to fail.

This commit fixes the issue by moving the BRIDGE variable definition to
where it's used in the networking functions.